Cleanup and make mypy happy
