+++ pyproject.toml
@@ -61,12 +61,12 @@ mcp = [
   "mcp>=1.5.0"
 ]
 
-tiny = [
+tinyagent = [
   "litellm>=1.0.0"
 ]
 
 all = [
-  "any-agent[google,langchain,llama_index,smolagents,openai,agno,mcp,tiny_agent,tiny]",
+  "any-agent[google,langchain,llama_index,smolagents,openai,agno,mcp,tinyagent]",
 ]
 
 docs = [
