+++ src/any_agent/config.py
@@ -12,7 +12,7 @@ class AgentFramework(str, Enum):
     OPENAI = auto()
     AGNO = auto()
     SMOLAGENTS = auto()
-    TINY = auto()
+    TINYAGENT = auto()
 
     @classmethod
     def from_string(cls, value: str | Self) -> Self:
