+++ src/any_agent/tools/__init__.py
@@ -7,7 +7,7 @@ from .mcp import (
     MCPServerBase,
     OpenAIMCPServer,
     SmolagentsMCPServer,
-    TinyMCPServer,
+    TinyAgentMCPServer,
     _get_mcp_server,
 )
 from .user_interaction import (
@@ -27,7 +27,7 @@ __all__ = [
     "MCPServerBase",
     "OpenAIMCPServer",
     "SmolagentsMCPServer",
-    "TinyMCPServer",
+    "TinyAgentMCPServer",
     "_get_mcp_server",
     "ask_user_verification",
     "search_web",
