+++ tests/integration/test_load_and_run_agent.py
@@ -42,7 +42,7 @@ def test_load_and_run_agent(agent_framework: AgentFramework, tmp_path: Path) -> 
     if agent_framework not in (
         AgentFramework.AGNO,
         AgentFramework.GOOGLE,
-        AgentFramework.TINY,
+        AgentFramework.TINYAGENT,
     ):
         assert traces.exists()
         assert agent_framework.name in str(next(traces.iterdir()).name)
